+++ app/src/main/java/com/fr1014/mycoludmusic/ui/block/RecommendPlayListBlock.kt
@@ -3,8 +3,10 @@ package com.fr1014.mycoludmusic.ui.block
 import android.content.Context
 import android.util.AttributeSet
 import android.view.LayoutInflater
+import android.view.View
 import android.widget.LinearLayout
 import android.widget.TextView
+import androidx.navigation.Navigation
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
@@ -13,6 +15,7 @@ import com.fr1014.frecyclerviewadapter.BaseViewHolder
 import com.fr1014.mycoludmusic.R
 import com.fr1014.mycoludmusic.data.entity.http.wangyiyun.PlayListResult
 import com.fr1014.mycoludmusic.databinding.BlockRecommendPlaylistBinding
+import com.fr1014.mycoludmusic.ui.home.toplist.PlayListDetailFragment
 
 class RecommendPlayListBlock @JvmOverloads constructor(
         context: Context, attrs: AttributeSet? = null, defStyleAttr: Int = 0
@@ -45,13 +48,26 @@ class RecommendPlayListBlock @JvmOverloads constructor(
     }
 }
 
-class RecommendPlayListAdapter(layoutResId: Int) : BaseAdapter<PlayListResult, BaseViewHolder>(layoutResId) {
+class RecommendPlayListAdapter(layoutResId: Int) : BaseAdapter<PlayListResult, BaseViewHolder>(layoutResId),BaseAdapter.OnItemClickListener{
 
     override fun convert(holder: BaseViewHolder, data: PlayListResult) {
         holder.getView<TextView>(R.id.tv_description).text = data.name
         Glide.with(holder.itemView)
                 .load(data.picUrl)
                 .into(holder.getView(R.id.iv_cover))
+        onItemClickListener = this
+        holder.addOnClickListener(R.id.item)
+    }
+
+    override fun onItemClick(adapter: BaseAdapter<*, *>, view: View, position: Int) {
+        when(view.id){
+            R.id.item -> {
+                val data = getData(position)
+                val bundle = PlayListDetailFragment.createBundle(data.id, data.name, data.picUrl)
+                Navigation.findNavController(view).navigate(R.id.playListDetailFragment,bundle)
+            }
+            else ->{}
+        }
     }
 
 }
